headdisplay: drawArwingHud 97.11->97.86 via (s16) cast on arwingHudAlpha clamp compares (#83c inverse) — keeps target's re-sign-extension (extsh) of the stored s16 before the >0xff / <0 checks

+++ src/main/dll/headdisplay.c
@@ -415,7 +415,7 @@ void drawArwingHud(void)
         if (arwingHudVisible != 0)
         {
             arwingHudAlpha = (int)(lbl_803E1FA0 * (f32)(u32)framesThisStep + arwingHudAlpha);
-            if (arwingHudAlpha > 0xff)
+            if ((s16)arwingHudAlpha > 0xff)
             {
                 arwingHudAlpha = 0xff;
             }
@@ -423,7 +423,7 @@ void drawArwingHud(void)
         else
         {
             arwingHudAlpha = (int) - (lbl_803E1FA0 * (f32)(u32)framesThisStep - arwingHudAlpha);
-            if (arwingHudAlpha < 0)
+            if ((s16)arwingHudAlpha < 0)
             {
                 arwingHudAlpha = 0;
             }
